command: テストコードを複数のファイルに分割する
コマンドごとに多くのテストケースがあり、1つのファイルが非常に大きく
なったため。
